Move battle SoundUtils into delegate.battle for cleaner dependencies
SoundUtils selects battle clips using the rules engine (Matches,
MustFightBattle.RetreatType, IDelegateBridge) but lived in org.triplea.sound.
Moving it to games.strategy.triplea.delegate.battle, alongside its callers,
strips every delegate and Unit import from the sound package so its
dependencies point only downward -- a step toward extracting org.triplea.sound
as a standalone module.
